Simplified user.db.ts
