+++ db/users/users.db.ts
@@ -1,6 +1,5 @@
 import { getAdminFirestore } from "@/lib/firebase/firebase-admin";
 import {
-  Notification,
   ContributionPlatform,
   UserRole,
   UserModel,
@@ -8,12 +7,11 @@ import {
 import { Timestamp } from "firebase-admin/firestore";
 import { DB_PATHS } from "../db-paths";
 import { DbNotFoundError, DbValidationError } from "../db.errors";
-import { normalizeNotificationDocument } from "../notifications/notifications.mapper";
+import { getUserNotificationsCollection } from "./notifications.db";
+import { serializeNotification } from "../notifications/notifications.mapper";
 import { normalizeUserDocument, serializeUser } from "./users.mapper";
 
 const db = getAdminFirestore();
-
-type NotificationStatusFilter = "READ" | "UNREAD" | "ALL";
 
 /**
  * Validates the GitHub username requirement for non-contributor roles.
@@ -234,10 +232,7 @@ export async function updateUserRoleAndTeamWithNotificationByUid(
   assertGithubUsernameForRole(githubUsername);
 
   const userDocRef = await getRequiredUserDocRefByUid(uid);
-
-  const notificationRef = userDocRef
-    .collection(DB_PATHS.USERS.NOTIFICATIONS_SUBCOLLECTION)
-    .doc();
+  const notificationRef = getUserNotificationsCollection(uid).doc();
   const batch = db.batch();
 
   batch.update(userDocRef, {
@@ -246,112 +241,12 @@ export async function updateUserRoleAndTeamWithNotificationByUid(
     githubUsername,
   });
   batch.set(notificationRef, {
-    message,
-    read: false,
-    createdAt: Timestamp.now(),
+    ...serializeNotification({
+      message,
+      read: false,
+      createdAt: new Date(),
+    }),
   });
 
   await batch.commit();
-}
-
-/**
- * Appends a notification to a user located by uid.
- *
- * @param uid The user id to update.
- * @param message The notification message to append.
- * @returns A promise that resolves when the notification has been written.
- */
-export async function appendUserNotificationByUid(
-  uid: string,
-  message: string,
-): Promise<void> {
-  const userDocRef = await getRequiredUserDocRefByUid(uid);
-
-  await appendNotificationByUserDocRef(userDocRef, {
-    message,
-    createdAt: new Date(),
-    read: false,
-  });
-}
-
-/**
- * Retrieves notifications for a user id, optionally filtered by read status.
- *
- * @param uid The user id to query by.
- * @param status The optional notification status filter.
- * @returns The normalized notifications sorted by creation time descending.
- */
-export async function getNotificationsByUid(
-  uid: string,
-  status: NotificationStatusFilter = "ALL",
-): Promise<Notification[]> {
-  const userDocRef = await getRequiredUserDocRefByUid(uid);
-
-  let query: FirebaseFirestore.Query = userDocRef.collection(
-    DB_PATHS.USERS.NOTIFICATIONS_SUBCOLLECTION,
-  );
-
-  if (status === "READ") {
-    query = query.where("read", "==", true);
-  } else if (status === "UNREAD") {
-    query = query.where("read", "==", false);
-  }
-
-  const snapshot = await query.orderBy("createdAt", "desc").get();
-
-  return snapshot.docs.map((doc) =>
-    normalizeNotificationDocument({
-      id: doc.id,
-      ...doc.data(),
-    }),
-  );
-}
-
-/**
- * Marks a notification as read for the user identified by uid.
- *
- * @param uid The user id to query by.
- * @param notificationId The notification document id to update.
- * @returns A promise that resolves when the notification has been marked as read.
- */
-export async function markNotificationAsReadByUid(
-  uid: string,
-  notificationId: string,
-): Promise<void> {
-  const userDocRef = await getRequiredUserDocRefByUid(uid);
-
-  const notificationRef = userDocRef
-    .collection(DB_PATHS.USERS.NOTIFICATIONS_SUBCOLLECTION)
-    .doc(notificationId);
-  const notificationSnap = await notificationRef.get();
-
-  if (!notificationSnap.exists) {
-    throw new DbNotFoundError("Notification");
-  }
-
-  await notificationRef.update({
-    read: true,
-  });
-}
-
-/**
- * Appends a notification document under an already-resolved user document reference.
- *
- * @param userDocRef The Firestore user document reference.
- * @param notification The notification payload to persist.
- * @returns A promise that resolves when the notification has been written.
- */
-async function appendNotificationByUserDocRef(
-  userDocRef: FirebaseFirestore.DocumentReference,
-  notification: Omit<Notification, "id">,
-): Promise<void> {
-  const notificationRef = userDocRef
-    .collection(DB_PATHS.USERS.NOTIFICATIONS_SUBCOLLECTION)
-    .doc();
-
-  await notificationRef.set({
-    message: notification.message,
-    read: notification.read,
-    createdAt: Timestamp.fromDate(notification.createdAt),
-  });
 }
